TAffineのLookAt系APIの仕様が変わりました。詳しくはAffine.hのコメントを見てください。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3911 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWLDHapticApp.cpp
@@ -96,7 +96,7 @@ void FWLDHapticApp::Step(){
 		bCalcPhys = false;
 	}
 	if (hapticcount < pdt/hdt) return;
-	hapticcount -= pdt/hdt;
+	hapticcount -= (int)(pdt/hdt);
 	bSync = true;
 	bCalcPhys = true;
 }
